[ASAP-1373] - Improve readability of call to cleanArray
